finish setup and mobile styling of question library
