+++ studyAi/src/app/library/questions/styles.ts
@@ -1,0 +1,14 @@
+const styles = {
+    layout: [
+        "py-16",
+        "px-5",
+        "sm:py-24",
+        "sm:px-16"
+    ].join(" "),
+    h1: [
+        "text-3xl",
+        "font-bold",
+    ].join(" "),
+}
+
+export default styles
